move group label tt tests to test_elementplot
