+++ holoviews/tests/plotting/bokeh/test_elementplot.py
@@ -2,6 +2,7 @@ import datetime as dt
 from unittest import SkipTest
 
 import numpy as np
+import pandas as pd
 import panel as pn
 import pytest
 from bokeh.document import Document
@@ -18,7 +19,7 @@ from bokeh.models import (
 
 from holoviews.core import DynamicMap, HoloMap, NdOverlay
 from holoviews.core.util import dt_to_int
-from holoviews.element import Curve, HeatMap, Image, Labels, Scatter
+from holoviews.element import Curve, HeatMap, Image, Labels, Points, Scatter
 from holoviews.plotting.util import process_cmap
 from holoviews.streams import PointDraw, Stream
 from holoviews.util import render
@@ -412,14 +413,6 @@ class TestElementPlot(LoggingComparisonTestCase, TestBokehPlot):
         self.assertIsInstance(toolbar.active_tap, tools.PointDrawTool)
         self.assertIsInstance(toolbar.active_drag, tools.PointDrawTool)
 
-    def test_hover_tooltip_update(self):
-        hmap = HoloMap({'a': Curve([1, 2, 3], vdims='a'), 'b': Curve([1, 2, 3], vdims='b')}).opts(
-            tools=['hover'])
-        plot = bokeh_renderer.get_plot(hmap)
-        self.assertEqual(plot.handles['hover'].tooltips, [('x', '@{x}'), ('a', '@{a}')])
-        plot.update(('b',))
-        self.assertEqual(plot.handles['hover'].tooltips, [('x', '@{x}'), ('b', '@{b}')])
-
     def test_categorical_dimension_values(self):
         curve = Curve([('C', 1), ('B', 3)]).redim.values(x=['A', 'B', 'C'])
         plot = bokeh_renderer.get_plot(curve)
@@ -792,6 +785,77 @@ class TestElementPlot(LoggingComparisonTestCase, TestBokehPlot):
         self.log_handler.assertContains(
             "WARNING", "cb model could not be"
         )
+
+    #################################################################
+    # Hover tooltip tests
+    #################################################################
+
+    def test_hover_tooltip_update(self):
+        hmap = HoloMap({'a': Curve([1, 2, 3], vdims='a'), 'b': Curve([1, 2, 3], vdims='b')}).opts(
+            tools=['hover'])
+        plot = bokeh_renderer.get_plot(hmap)
+        self.assertEqual(plot.handles['hover'].tooltips, [('x', '@{x}'), ('a', '@{a}')])
+        plot.update(('b',))
+        self.assertEqual(plot.handles['hover'].tooltips, [('x', '@{x}'), ('b', '@{b}')])
+
+    def test_group_label_overridden_by_data_dims(self):
+        """test if the group and label args will be overridden by data dims of the same name"""
+        group = "A"
+        label = "B"
+        data = pd.DataFrame(dict(group=range(4), Label=range(4)))
+        obj = Curve(data, group=group, label=label).opts(tools=["hover"])
+        tooltips = [
+            ("group", "@{group}"),
+            ("Label", "@{Label}"),
+        ]
+        self._test_hover_info(obj, tooltips, "nearest")
+
+    def test_label_tooltip_from_arg(self):
+        """test if the label arg will be added to the tooltips if not present in the data"""
+        data = np.random.rand(10, 2)
+        label = "B"
+        obj = Points(data, kdims=["Group", "Dim2"], label=label).opts(tools=["hover"])
+        tooltips = [
+            ("Label", label),
+            ("Group", "@{Group}"),
+            ("Dim2", "@{Dim2}"),
+        ]
+        self._test_hover_info(obj, tooltips, "nearest")
+
+    def test_group_not_overridden_by_similar_data_dim(self):
+        """test if a data dim that has a matching part only will override group arg"""
+        data = np.random.rand(10, 2)
+        group = "A"
+        label = "B"
+        obj = Curve(data, kdims=["GroupA"], group=group, label=label).opts(
+            tools=["hover"]
+        )
+        tooltips = [
+            ("Group", group),
+            ("Label", label),
+            ("GroupA", "@{GroupA}"),
+            ('y', '@{y}')
+        ]
+        self._test_hover_info(obj, tooltips, "nearest")
+
+    def test_group_label_batched(self):
+        """test if the group as a container dim will override the group arg on individual elements"""
+        obj = NdOverlay(
+            {
+                i: Points(
+                    [np.random.rand(10, 2)],
+                    kdims=["Label", "Dim2"],
+                    group="arg_test",
+                    label="arg_test",
+                )
+                for i in range(5)
+            },
+            kdims=["group"],
+        )
+        opts = {"Points": {"tools": ["hover"]}, "NdOverlay": {"legend_limit": 0}}
+        obj = obj.opts(opts)
+        tooltips = [('group', '@{group}'), ('Label', '@{Label}'), ('Dim2', '@{Dim2}')]
+        self._test_hover_info(obj, tooltips)
 
 class TestColorbarPlot(LoggingComparisonTestCase, TestBokehPlot):
